move some things into components
